requirements.txt
aiodns==1.1.1
aiohttp==3.4.4
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
ccxt==1.17.371
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3.1
dateparser==0.7.0
hyperlink==18.0.0
idna==2.7
idna-ssl==1.1.0
incremental==17.5.0
multidict==4.4.2
numpy==1.15.2
pdb-clone==1.10.1
psycopg2-binary==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycares==2.3.0
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
python-binance==0.7.0
python-dateutil==2.7.3
pytz==2018.5
pyzmq==17.1.2
regex==2018.8.29
requests==2.19.1
requests-toolbelt==0.8.0
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.2.12
Twisted==18.7.0
txaio==18.8.1
tzlocal==1.5.1
urllib3==1.23
yarl==1.1.0
zmq==0.0.0
zope.interface==4.5.0
absl-py
